update npm packages, fix stage component build erro
