+++ package.json
@@ -80,7 +80,7 @@
     "cookie-parser": "1.4.5",
     "cookie-session": "2.0.0-beta.3",
     "cross-env": "7.0.2",
-    "drag-drop-stage-component": "1.0.24",
+    "drag-drop-stage-component": "1.0.27",
     "express": "4.17.1",
     "express-force-ssl": "0.3.2",
     "font-awesome": "4.7.0",
@@ -103,20 +103,19 @@
     "wysihtml": "0.6.0-beta1"
   },
   "devDependencies": {
-    "@babel/preset-env": "7.10.2",
-    "@babel/preset-typescript": "7.10.1",
-    "@types/express": "^4.17.6",
-    "@types/express-session": "^1.17.0",
-    "@types/jest": "26.0.0",
+    "@babel/preset-env": "7.10.4",
+    "@babel/preset-typescript": "7.10.4",
+    "@types/express": "4.17.6",
+    "@types/express-session": "1.17.0",
+    "@types/jest": "26.0.3",
     "@types/jsdom": "16.2.3",
-    "@types/node": "14.0.13",
-    "@types/redux": "3.6.31",
-    "babel-jest": "26.0.1",
-    "css-loader": "3.5.3",
-    "detect-browser": "5.1.0",
-    "eslint": "7.2.0",
-    "file-loader": "^6.0.0",
-    "jest": "26.0.1",
+    "@types/node": "14.0.14",
+    "babel-jest": "26.1.0",
+    "css-loader": "3.6.0",
+    "detect-browser": "5.1.1",
+    "eslint": "7.3.1",
+    "file-loader": "6.0.0",
+    "jest": "26.1.0",
     "less": "3.11.3",
     "monaco-editor": "0.20.0",
     "monaco-editor-webpack-plugin": "1.9.0",
@@ -124,12 +123,12 @@
     "npm-watch": "0.6.0",
     "pug-cli": "1.0.0-alpha6",
     "style-loader": "1.2.1",
-    "ts-jest": "26.1.0",
+    "ts-jest": "26.1.1",
     "ts-loader": "7.0.5",
     "tslint": "6.1.2",
     "typescript": "3.9.5",
     "webpack": "4.43.0",
-    "webpack-cli": "3.3.11"
+    "webpack-cli": "3.3.12"
   },
   "repository": {
     "type": "git",
