+++ tsconfig.json
@@ -7,7 +7,11 @@
     "sourceMap": true,
     "module": "esnext",
     "removeComments": false,
-    "typeRoots": ["src/ts/types"],
+    "typeRoots": [
+      "src/ts/types",
+      "node_modules/@types"
+    ],
+    "moduleResolution": "node",
     "esModuleInterop": true
   },
   "include": [
